Use 'pdfenconding=auto' instead of 'unicode' for package 'hyperref'.
The package 'hyperref' is quite troublesome and should be avoided.
Nonetheless, if used, we must configure everything as correctly as
possible. Until now, we were using the 'unicode' option when loading
'hyperref'. This will encode every PDF string as Unicode (PDF Strings
are used for bookmarks, for instance). However, this is not often
required. An alternative is to let LaTeX choose unicode if the string
has non-ASCII characters. The option 'pdfencoding=auto' does just this.

References:
https://tex.stackexchange.com/questions/24445/hyperref-lualatex-and-unicode-bookmarks-issue-garbled-page-numbers-in-ar-for-l
https://groups.google.com/forum/#!topic/comp.text.tex/6fYmcVMbSbQ/discussion